backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
sqlparse==0.5.1

asgiref==3.8.1
certifi==2024.8.30
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.0
pytest==8.3.3
pytest-django==4.9.0
python-telegram-bot==21.7
sniffio==1.3.1
sqlparse==0.5.1
